+++ src/comp-schema.ts
@@ -68,6 +68,8 @@ export function genMethodSchema(o = Object.create(null)): MethodSchema {
 }
 
 export const SCHEME_KEY = "__schema__";
+export type ISchema = { [SCHEME_KEY]: Schema };
+
 export function getSchemaByPrototype(prototype: any): Schema | null {
     if (prototype.hasOwnProperty(SCHEME_KEY)) {
         return (prototype as any)[SCHEME_KEY];
