+++ src/macro.ts
@@ -1,2 +1,4 @@
 export const NULL_NUM = -1;
 export const NULL_STR = "";
+export const NONE_CONTAINER = 0;
+export const ARR_CONTAINER = 1;
